+++ Plugins/ElvUI.lua
@@ -249,11 +249,13 @@ local function SkinAchievementSummary(frame, engine, skins)
     frame.Achievements.Header.Texture:Hide();
 	frame.Categories.Header.Texture:Hide();
 
-    local numChildren = frame:GetNumChildren();
-    for i = 1, numChildren do
-        local child = select(i, frame:GetChildren());
-        if child and not child:GetName() then
-            child:SetBackdrop();
+    if not addon.IsWrathClassic then
+        local numChildren = frame:GetNumChildren();
+        for i = 1, numChildren do
+            local child = select(i, frame:GetChildren());
+            if child and not child:GetName() then
+                child:SetBackdrop();
+            end
         end
     end
 
@@ -432,10 +434,15 @@ local function SkinHeader()
         AchievementFrameHeaderPoints:ClearAllPoints();
 	    AchievementFrameHeaderPoints:Point('CENTER', AchievementFrameHeaderPointBorder);
     end);
+    if addon.IsWrathClassic then
+        AchievementFrameHeaderLeftDDLInset:SetAlpha(0);
+    end
 end
 
 local function ReskinBlizzard(skins)
-    SkinSearchBoxFrame(AchievementFrame.searchBox, AchievementFrameAchievements, skins);
+    if not addon.IsWrathClassic then
+        SkinSearchBoxFrame(AchievementFrame.SearchBox, AchievementFrameAchievements, skins);
+    end
     AchievementFrameFilterDropDown:ClearAllPoints();
 	AchievementFrameFilterDropDown:Point('TOPLEFT', AchievementFrameAchievements, 'TOPLEFT', -18, 26);
     AchievementFrameFilterDropDown:Size(AchievementFrameFilterDropDown:GetWidth(), AchievementFrameFilterDropDown:GetHeight() - 1);