+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 44.2 - 2022-10-14
+### Fixed
+- Temporary obtainable account wide achievements should no longer display the Earned By section
+
+### Wrath Classic
+- Fixed ElvUI compatibility
+
 ## 44.1 - 2022-10-13
 ### Fixed
 - Silent World Map Buttons